feat: Add date-fns library for date formatting
Add the date-fns library to the project dependencies for date formatting. This will be used in the my-storage page to display the last modified date of folders. The library is added to the package.json and pnpm-lock.yaml files. The format function from date-fns is imported in the page.tsx file and used to format the last modified date of folders. The commit also includes removing unused code and updating the .gitignore file to include .env.